DM-126 Organized and refactored Sleeper League Components
