+++ frontend/src/feature/search/sleeper/components/SleeperLeaguesList.tsx
@@ -6,7 +6,7 @@ import { useAppSelector } from "@app/hooks";
 import { DisplayLeaguesList } from "@components/DisplayLeaguesList";
 import SelectSeasonDropDown from "@components/SelectSeasonDropDown";
 
-import useSaveSleeperLeague from "@feature/leagues/hooks/league/useSaveTeam";
+import useSaveSleeperLeague from "@feature/leagues/sleeper/hooks/team/useSaveTeam";
 import useDeleteLeague from "@feature/search/sleeper/hooks/useDeleteLeague";
 import useGetUserLeaguesSleeper from "@feature/search/sleeper/hooks/useGetUserLeaguesSleeper";
 import { setSubmit } from "@feature/search/sleeper/sleeperSearchSlice";
